Detect April Tags
This was the only program that we tried so far that can detect april tags and their ID. This was just a sample that was already included in FtcRobotController, except we slightly modified it by changing file location(line 30), removing @disabled(88), adding variable arm(105), and changing the hardware map(125-127). The original sample was called, "RobotAutoDriveToAprilTagTank.java". We where about to test the sample titled, "ConceptAprilTagEasy.java", but we didn't get to that yet